If a list contains maps with computed entries, then mark the list as computed.
This fixes a problem for fields of type `schema.TypeSet`, whose elements are of type `schema.TypeMap`, whose elements are computed.

For example, consider the following schema...

```
	"maps": &schema.Schema{
		Type:        schema.TypeSet,
		Set:         mapHash,
		MinItems:    0,

		Elem: &schema.Schema{
			Type:        schema.TypeMap,
			MinItems:    1,
			MaxItems:    1,
			Elem: &schema.Schema{
				Type:        schema.TypeString,
			},
		},
		Optional: true,
	},
```

Note that the `mapHash` function determines the unique ID of an item (a map) so that a proper set can be built.

Without this change, the `mapHash` function will be passed `nil` for computed maps (since field_reader.ReadField returns null for maps with computed entries).
A proper hash for `nil` cannot be accurately determined.  You could theoretically always return the same value, but that breaks downstream logic that assumes each element will have a unique hash.

With this change, since the list is marked as computed, `mapHash` is not invoked for each element.

This is causing Pryz/terraform-provider-ldap#6